feat(unity): Phase 2 play loop + living-city life layer scaffolds
Wire Forge.SimCore SimHost into Unity with roads, budget/research/Herald panels,
CMJR save/load, GLTF building instancing, and cosmetic life systems (vehicles,
pedestrians, growth staging, service/edge overlays, ambient time-of-day).
Add LifeSimMath, Population L2 bridge, and ecosystem vision docs for v1.1–v2.5.

+++ src/Forge.Game/Simulation/TradeSystem.cs
@@ -316,7 +316,7 @@ public sealed class TradeSystem
         // Small random walk: each good drifts -2% to +2% per month
         for (int g = 0; g < EconomySystem.GoodCount; g++)
         {
-            float drift = (_rng.NextSingle() - 0.5f) * 0.04f; // -2% to +2%
+            float drift = ((float)_rng.NextDouble() - 0.5f) * 0.04f; // -2% to +2%
             GlobalPrices[g] *= (1f + drift);
 
             // Mean-revert toward base price: 5% pull per month
